+++ piplicenses_lib/__init__.py
@@ -222,6 +222,27 @@ def _get_other_files(package: Distribution, existing_files: Iterable[tuple[str, 
         yield other_file_str, read_file(other_file_resolved)
 
 
+def _get_sboms(package: Distribution) -> Generator[tuple[str, str]]:
+    """
+    Get PEP 770 compliant files.
+
+    References:
+        * https://peps.python.org/pep-0770/
+        * https://packaging.python.org/en/latest/specifications/binary-distribution-format/#the-dist-info-sboms-directory
+
+    :param package: The package to work on.
+    :return: A list/generator of tuples holding the file path and corresponding file content.
+    """
+    if not isinstance(package, PathDistribution):
+        return
+    path = Path(str(package._path), "sboms")
+    if not path.is_dir():
+        # No SBOMs
+        return
+    for inner_path in sorted(path.iterdir()):
+        yield str(inner_path), read_file(inner_path)
+
+
 @dataclass
 class PackageInfo:
     """
@@ -293,6 +314,11 @@ class PackageInfo:
     List of other licensing-related files and their contents.
     """
 
+    sboms: list[tuple[str, str]] = dataclass_field(default_factory=list)
+    """
+    List of SBOM files and their contents.
+    """
+
     requirements: set[str] = dataclass_field(default_factory=set)
     """
     Collection of all declared (direct) requirements.
@@ -353,6 +379,22 @@ class PackageInfo:
         for entry in self.others:
             yield entry[1]
 
+    @property
+    def sbom_files(self) -> Iterator[str]:
+        """
+        List of SBOM file paths.
+        """
+        for entry in self.sboms:
+            yield entry[0]
+
+    @property
+    def sbom_texts(self) -> Iterator[str]:
+        """
+        List of SBOM contents.
+        """
+        for entry in self.sboms:
+            yield entry[1]
+
 
 def get_package_info(
         package: Distribution, include_files: bool = True, normalize_name: bool = True,
@@ -371,10 +413,12 @@ def get_package_info(
         ))
         notice_files = list(get_package_included_files(package, "NOTICE.*"))
         other_files = list(_get_other_files(package, itertools.chain(license_files, notice_files)))
+        sboms = list(_get_sboms(package))
     else:
         license_files = []
         notice_files = []
         other_files = []
+        sboms = []
 
     name = package.metadata["name"]
     if normalize_name:
@@ -385,6 +429,7 @@ def get_package_info(
         licenses=license_files,
         notices=notice_files,
         others=other_files,
+        sboms=sboms,
         requirements=set(package.requires or []),
         distribution=package,
     )